Split a few macro helpers out into separate files
