+++ crates/test-assembly/crates/test_msg_send_error/expected/apple-aarch64.s
@@ -1,6 +1,6 @@
 	.section	__TEXT,__text,regular,pure_instructions
 	.p2align	2
-SYM(objc2[CRATE_ID]::__macro_helpers::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0):
+SYM(objc2[CRATE_ID]::__macro_helpers::msg_send_id::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0):
 	stp	x20, x19, [sp, #-32]!
 	stp	x29, x30, [sp, #16]
 	add	x29, sp, #16
@@ -85,7 +85,7 @@ Lloh6:
 	adrp	x1, l_anon.[ID].4@PAGE
 Lloh7:
 	add	x1, x1, l_anon.[ID].4@PAGEOFF
-	bl	SYM(objc2[CRATE_ID]::__macro_helpers::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
+	bl	SYM(objc2[CRATE_ID]::__macro_helpers::msg_send_id::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
 	mov	x1, x0
 	mov	w0, #1
 	ldp	x29, x30, [sp, #16]
@@ -114,7 +114,7 @@ Lloh8:
 	adrp	x1, l_anon.[ID].5@PAGE
 Lloh9:
 	add	x1, x1, l_anon.[ID].5@PAGEOFF
-	bl	SYM(objc2[CRATE_ID]::__macro_helpers::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
+	bl	SYM(objc2[CRATE_ID]::__macro_helpers::msg_send_id::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
 	mov	x1, x0
 	mov	w0, #1
 	ldp	x29, x30, [sp, #16]
@@ -143,7 +143,7 @@ Lloh10:
 	adrp	x1, l_anon.[ID].6@PAGE
 Lloh11:
 	add	x1, x1, l_anon.[ID].6@PAGEOFF
-	bl	SYM(objc2[CRATE_ID]::__macro_helpers::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
+	bl	SYM(objc2[CRATE_ID]::__macro_helpers::msg_send_id::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
 	mov	x1, x0
 	mov	w0, #1
 	ldp	x29, x30, [sp, #16]
@@ -172,7 +172,7 @@ Lloh12:
 	adrp	x1, l_anon.[ID].7@PAGE
 Lloh13:
 	add	x1, x1, l_anon.[ID].7@PAGEOFF
-	bl	SYM(objc2[CRATE_ID]::__macro_helpers::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
+	bl	SYM(objc2[CRATE_ID]::__macro_helpers::msg_send_id::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
 	mov	x1, x0
 	mov	w0, #1
 	ldp	x29, x30, [sp, #16]
@@ -205,7 +205,7 @@ Lloh14:
 	adrp	x1, l_anon.[ID].8@PAGE
 Lloh15:
 	add	x1, x1, l_anon.[ID].8@PAGEOFF
-	bl	SYM(objc2[CRATE_ID]::__macro_helpers::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
+	bl	SYM(objc2[CRATE_ID]::__macro_helpers::msg_send_id::encountered_error::<objc2[CRATE_ID]::runtime::AnyObject>, 0)
 	mov	x1, x0
 	mov	w0, #1
 	ldp	x29, x30, [sp, #16]
