+++ crates/test-assembly/crates/test_msg_send_id/expected/apple-aarch64.s
@@ -24,7 +24,7 @@ Lloh1:
 	add	x2, x2, l_anon.[ID].1@PAGEOFF
 	mov	x0, x20
 	mov	x1, x19
-	bl	SYM(<objc2::__macro_helpers::RetainSemantics<1_u8> as objc2::__macro_helpers::MsgSendIdFailed>::failed::GENERATED_ID, 0)
+	bl	SYM(<objc2::__macro_helpers::method_family::RetainSemantics<1_u8> as objc2::__macro_helpers::msg_send_id::MsgSendIdFailed>::failed::GENERATED_ID, 0)
 	.loh AdrpAdd	Lloh0, Lloh1
 
 	.globl	_handle_alloc
@@ -52,7 +52,7 @@ Lloh3:
 	add	x2, x2, l_anon.[ID].2@PAGEOFF
 	mov	x0, x20
 	mov	x1, x19
-	bl	SYM(<objc2::__macro_helpers::RetainSemantics<2_u8> as objc2::__macro_helpers::MsgSendIdFailed>::failed::GENERATED_ID, 0)
+	bl	SYM(<objc2::__macro_helpers::method_family::RetainSemantics<2_u8> as objc2::__macro_helpers::msg_send_id::MsgSendIdFailed>::failed::GENERATED_ID, 0)
 	.loh AdrpAdd	Lloh2, Lloh3
 
 	.globl	_handle_init
@@ -80,7 +80,7 @@ Lloh5:
 	add	x2, x2, l_anon.[ID].3@PAGEOFF
 	mov	x0, x20
 	mov	x1, x19
-	bl	SYM(<objc2::__macro_helpers::RetainSemantics<3_u8> as objc2::__macro_helpers::MsgSendIdFailed>::failed::GENERATED_ID, 0)
+	bl	SYM(<objc2::__macro_helpers::method_family::RetainSemantics<3_u8> as objc2::__macro_helpers::msg_send_id::MsgSendIdFailed>::failed::GENERATED_ID, 0)
 	.loh AdrpAdd	Lloh4, Lloh5
 
 	.globl	_handle_alloc_init
@@ -138,7 +138,7 @@ Lloh6:
 	adrp	x0, l_anon.[ID].4@PAGE
 Lloh7:
 	add	x0, x0, l_anon.[ID].4@PAGEOFF
-	bl	SYM(<objc2::__macro_helpers::RetainSemantics<4_u8> as objc2::__macro_helpers::MsgSendIdFailed>::failed::GENERATED_ID, 0)
+	bl	SYM(<objc2::__macro_helpers::method_family::RetainSemantics<4_u8> as objc2::__macro_helpers::msg_send_id::MsgSendIdFailed>::failed::GENERATED_ID, 0)
 	.loh AdrpAdd	Lloh6, Lloh7
 
 	.globl	_handle_autoreleased
@@ -177,7 +177,7 @@ Lloh9:
 	add	x2, x2, l_anon.[ID].5@PAGEOFF
 	mov	x0, x20
 	mov	x1, x19
-	bl	SYM(<objc2::__macro_helpers::RetainSemantics<5_u8> as objc2::__macro_helpers::MsgSendIdFailed>::failed::GENERATED_ID, 0)
+	bl	SYM(<objc2::__macro_helpers::method_family::RetainSemantics<5_u8> as objc2::__macro_helpers::msg_send_id::MsgSendIdFailed>::failed::GENERATED_ID, 0)
 	.loh AdrpAdd	Lloh8, Lloh9
 
 	.globl	_handle_with_out_param
