+++ crates/test-assembly/crates/test_msg_send_id/expected/gnustep-x86_64.s
@@ -44,7 +44,7 @@ handle_new_fallible:
 	lea	rdx, [rip + .Lanon.[ID].1]
 	mov	rdi, r14
 	mov	rsi, rbx
-	call	qword ptr [rip + SYM(<objc2::__macro_helpers::RetainSemantics<1_u8> as objc2::__macro_helpers::MsgSendIdFailed>::failed::GENERATED_ID, 0)@GOTPCREL]
+	call	qword ptr [rip + SYM(<objc2::__macro_helpers::method_family::RetainSemantics<1_u8> as objc2::__macro_helpers::msg_send_id::MsgSendIdFailed>::failed::GENERATED_ID, 0)@GOTPCREL]
 	ud2
 .Lfunc_end1:
 	.size	handle_new_fallible, .Lfunc_end1-handle_new_fallible
@@ -93,7 +93,7 @@ handle_alloc_fallible:
 	lea	rdx, [rip + .Lanon.[ID].2]
 	mov	rdi, r14
 	mov	rsi, rbx
-	call	qword ptr [rip + SYM(<objc2::__macro_helpers::RetainSemantics<2_u8> as objc2::__macro_helpers::MsgSendIdFailed>::failed::GENERATED_ID, 0)@GOTPCREL]
+	call	qword ptr [rip + SYM(<objc2::__macro_helpers::method_family::RetainSemantics<2_u8> as objc2::__macro_helpers::msg_send_id::MsgSendIdFailed>::failed::GENERATED_ID, 0)@GOTPCREL]
 	ud2
 .Lfunc_end3:
 	.size	handle_alloc_fallible, .Lfunc_end3-handle_alloc_fallible
@@ -151,7 +151,7 @@ handle_init_fallible:
 	lea	rdx, [rip + .Lanon.[ID].3]
 	mov	rdi, r14
 	mov	rsi, rbx
-	call	qword ptr [rip + SYM(<objc2::__macro_helpers::RetainSemantics<3_u8> as objc2::__macro_helpers::MsgSendIdFailed>::failed::GENERATED_ID, 0)@GOTPCREL]
+	call	qword ptr [rip + SYM(<objc2::__macro_helpers::method_family::RetainSemantics<3_u8> as objc2::__macro_helpers::msg_send_id::MsgSendIdFailed>::failed::GENERATED_ID, 0)@GOTPCREL]
 	ud2
 .Lfunc_end5:
 	.size	handle_init_fallible, .Lfunc_end5-handle_init_fallible
@@ -291,7 +291,7 @@ handle_copy_fallible:
 	ret
 .LBB10_1:
 	lea	rdi, [rip + .Lanon.[ID].4]
-	call	qword ptr [rip + SYM(<objc2::__macro_helpers::RetainSemantics<4_u8> as objc2::__macro_helpers::MsgSendIdFailed>::failed::GENERATED_ID, 0)@GOTPCREL]
+	call	qword ptr [rip + SYM(<objc2::__macro_helpers::method_family::RetainSemantics<4_u8> as objc2::__macro_helpers::msg_send_id::MsgSendIdFailed>::failed::GENERATED_ID, 0)@GOTPCREL]
 	ud2
 .Lfunc_end10:
 	.size	handle_copy_fallible, .Lfunc_end10-handle_copy_fallible
@@ -344,7 +344,7 @@ handle_autoreleased_fallible:
 	lea	rdx, [rip + .Lanon.[ID].5]
 	mov	rdi, r14
 	mov	rsi, rbx
-	call	qword ptr [rip + SYM(<objc2::__macro_helpers::RetainSemantics<5_u8> as objc2::__macro_helpers::MsgSendIdFailed>::failed::GENERATED_ID, 0)@GOTPCREL]
+	call	qword ptr [rip + SYM(<objc2::__macro_helpers::method_family::RetainSemantics<5_u8> as objc2::__macro_helpers::msg_send_id::MsgSendIdFailed>::failed::GENERATED_ID, 0)@GOTPCREL]
 	ud2
 .Lfunc_end12:
 	.size	handle_autoreleased_fallible, .Lfunc_end12-handle_autoreleased_fallible
