+++ crates/test-ui/ui/extern_methods_invalid_type.stderr
@@ -48,7 +48,7 @@ error[E0277]: the trait bound `i32: MaybeUnwrap` is not satisfied
             Option<Id<T>>
             Option<Allocated<T>>
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/msg_send_id.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                       ^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
@@ -71,7 +71,7 @@ error[E0308]: mismatched types
   = note:           expected enum `Option<Allocated<_>>`
           found mutable reference `&mut MyObject`
 note: associated function defined here
- --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/msg_send_id.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |               ^^^^^^^^^^^^^^^
@@ -127,7 +127,7 @@ error[E0277]: the trait bound `Result<Id<MyObject>, Id<NSObject>>: MaybeUnwrap` 
             Option<Id<T>>
             Option<Allocated<T>>
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/msg_send_id.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                       ^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
