+++ crates/test-ui/ui/msg_send_id_invalid_receiver.stderr
@@ -10,7 +10,7 @@ error[E0308]: mismatched types
   = note: expected reference `&AnyClass`
              found reference `&NSObject`
 note: associated function defined here
- --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/msg_send_id.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |               ^^^^^^^^^^^^^^^
@@ -27,7 +27,7 @@ error[E0308]: mismatched types
   = note:   expected enum `Option<Allocated<_>>`
           found reference `&NSObject`
 note: associated function defined here
- --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/msg_send_id.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |               ^^^^^^^^^^^^^^^
@@ -44,7 +44,7 @@ error[E0308]: mismatched types
   = note:   expected enum `Option<Allocated<_>>`
           found reference `&AnyClass`
 note: associated function defined here
- --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/msg_send_id.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |               ^^^^^^^^^^^^^^^
@@ -61,7 +61,7 @@ error[E0308]: mismatched types
   = note: expected enum `Option<Allocated<_>>`
            found struct `Id<NSObject>`
 note: associated function defined here
- --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/msg_send_id.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |               ^^^^^^^^^^^^^^^
@@ -78,7 +78,7 @@ error[E0308]: mismatched types
   = note: expected enum `Option<Allocated<_>>`
              found enum `Option<Id<NSObject>>`
 note: associated function defined here
- --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/msg_send_id.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |               ^^^^^^^^^^^^^^^
