+++ crates/test-ui/ui/msg_send_underspecified_error.stderr
@@ -37,7 +37,7 @@ error[E0283]: type annotations needed
             where T: RefEncode, T: ?Sized;
   = note: required for `*mut *mut _` to implement `Encode`
 note: required by a bound in `send_message_id_error`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/msg_send_id.rs
   |
   |     unsafe fn send_message_id_error<A, E>(obj: T, sel: Sel, args: A) -> Result<U, Id<E>>
   |               --------------------- required by a bound in this associated function
@@ -58,7 +58,7 @@ error[E0283]: type annotations needed
             where T: RefEncode, T: ?Sized;
   = note: required for `*mut *mut _` to implement `Encode`
 note: required by a bound in `send_message_id_error`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/msg_send_id.rs
   |
   |     unsafe fn send_message_id_error<A, E>(obj: T, sel: Sel, args: A) -> Result<U, Id<E>>
   |               --------------------- required by a bound in this associated function
@@ -79,7 +79,7 @@ error[E0283]: type annotations needed
             where T: RefEncode, T: ?Sized;
   = note: required for `*mut *mut _` to implement `Encode`
 note: required by a bound in `send_message_id_error`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/msg_send_id.rs
   |
   |     unsafe fn send_message_id_error<A, E>(obj: T, sel: Sel, args: A) -> Result<U, Id<E>>
   |               --------------------- required by a bound in this associated function
